Remy going slowly insane, gonna clean  up code for fun
